Least Squares Fitting

These functions are specific to least squares fitting applications.

General Application

Least squares fitting using ordinary or generalized least squares is supported.

ExtraStats.fittest — Function

fittest(x, y; fittype = :OLS, fitmodel = nothing, p0 = [0.1, 0.1], figshow = true)

Perform a least squares fit with dependant variable y and independant x, test the significance of the trend, and create a plot to evaluate the residuals.

Default fitmodel is linear, and fittype ordinary least squares fit, :OLS, but optionally perform generalized least squares with fittype = :GLS.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> fittest(x, y; fittype = :OLS, fitmodel = fmodel, p0 = [0.1, 0.1], figshow = true)
(LsqFit.LsqFitResult{Vector{Float64}, [...]}
------   --------------   --------------
, Float64[]), Test for nonzero correlation
----------------------------
Population details:
    parameter of interest:   Correlation
    value under h_0:         0.0
    point estimate:          0.0611079
    95% confidence interval: (-0.0008886, 0.1226)

Test summary:
    outcome with 95% confidence: fail to reject h_0
    two-sided p-value:           0.0534

Details:
    number of observations:          1000
    number of conditional variables: 0
    t-statistic:                     1.93408
    degrees of freedom:              998
, Scene (1200px, 750px):
  0 Plots
  15 Child Scenes:
    ├ Scene (1200px, 750px)
    [...]

ExtraStats.OLS — Function

OLS(x, y; fitmodel = nothing, p0 = [0.1, 0.1])

Perform an ordinary least squares fit with dependant variable y and independant x.

Default fitmodel is linear.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> OLS(x, y; fitmodel = fmodel, p0 = [0.1, 0.1])
LsqFit.LsqFitResult{Vector{Float64}, [...]}

ExtraStats.GLS — Function

GLS(x, y; K = nothing, fitmodel = nothing, p0 = [0.1, 0.1])

Perform a generalized least squares fit with dependant variable y and independant x.

Default fitmodel is linear.

Weight PrecisionMatrix K is estimated from the residuals of OLS by default.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> GLS(x, y; fitmodel = fmodel, p0 = [0.1, 0.1])
LsqFit.LsqFitResult{Vector{Float64}, [...]}

ExtraStats.sqcovar — Function

sqcovar(X)

Calculate the Hermitian square root auto covariance matrix of input vector X.

Fit Evaluation

Several functions for evaluating the goodness-of-fit of a given fit are provided.

ExtraStats.CalculateRMSE — Function

CalculateRMSE(e)

Calculate the Root-Mean-Square-Error of the vector of residuals e.

ExtraStats.r² — Function

r²(data, e)

Estimate the coefficient of determination of a fit to data with residuals e.

ExtraStats.χ²Test — Function

χ²Test(x, CDF, p; nbins=500, siglevel = 0.99)

Perform a χ² goodness-of-fit test of the null hypothesis that the data x come from a distribution with cumulative distribution function cdf with parameters p.

Optionally specify the number of bins nbins to separate x into, and the significance level siglevel to test against.

ExtraStats.KuiperTest — Function

KuiperTest(x, Fₕ)

Perform Kuiper's goodness-of-fit test of the null hypothesis that the data x come from CDF Fₕ.

Linear Fit Uncertainty

Specifically for linear fits, uncertainty on the slope and intercept can easily be estimated.

ExtraStats.linear_uncertainty — Function

linear_uncertainty(xi, yi, yhat, γ)

Calculate the confidence intervals at the γ level on the fit parameters of a linear regression on indpendent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.SE_beta — Function

SE_beta(xi, yi, yhat)

Calculate the standard error of the slope of a linear regression with independent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.SE_alpha — Function

SE_alpha(xi, yi, yhat)

Calculate the standard error of the intercept of a linear regression with independent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.uncertainty_lines — Function

uncertainty_lines(β_est, α_est, x, t_val, xi, yi, yhat)

Create the upper and lower confidence bound lines of a linear regression.

...

Arguments

β_est: best fit estimate of the slope.

α_est: best fit estimate of the intercept.

x: independent variable values to calculate bounds at.

xi: independent variable observations.

yi: dependent variable observations.

yhat: predicted values of the dependent variable.

γ: confidence level (i.e. γ-th percentile).

...