Statistical Distributions

The mathematical definitions of several common (in Oceanography) probability density functions and their cumulative distribution functions are defined here for general use.

General Application

In addition, some generally-useful functions related to statistical distributions are provided.

ExtraStats.find_pdf_mode — Function

find_pdf_mode(pdf, x0; lr=nothing, up=nothing)

Estimate the mode of a probability density function pdf with initial guess x0.

ExtraStats.find_HDI — Function

find_HDI(pdf, samples; target_mass = 0.68)

Find the Highest Density Interval (Hyndman 1996) for the analytic probability density function pdf.

ExtraStats.transform_pdf — Function

transform_pdf(pdf, g, g⁻¹)

Complete a change of variables on an analytic probability density function pdf using the inverse g⁻¹ of the transformation function.

Normal Functions

ExtraStats.gaussian — Function

gaussian(x, A, μ, σ)

Calculate the value of a Gaussian function with amplitude A, centre μ, and standard deviation σ at x.

gaussian(x, p)

Calculate the value of a Gaussian function with parameter vector p at x.

ExtraStats.normal — Function

normal(x, μ, σ)

Calculate the value of a Normal function with centre μ and standard deviation σ at x.

normal(x, p)

Calculate the value of a Normal function with parameter vector p at x.

ExtraStats.∫Normaldx — Function

∫Normaldx(x)

Calculate the value of the integral of a Normal function at x.

ExtraStats.NormalCDF — Function

NormalCDF(x, μ, σ)

Calculate the CDF of a Normal pdf with centre μ and standard deviation σ at x.

Log-Normal Functions

ExtraStats.lognormal — Function

lognormal(x, μ, σ)

Calculate the value of a Log-Normal function with log-centre μ and log-standard deviation σ at x.

lognormal(x, p)

Calculate the value of a Log-Normal function with parameter vector p at x.

ExtraStats.LogNormalCDF — Function

LogNormalCDF(x, μ, σ)

Calculate the CDF of a Log-Normal pdf with centre μ and standard deviation σ at x.

Skew-Normal Functions

ExtraStats.skewgaussian — Function

skewgaussian(x, A, λ, ω, ξ)

Calculate the value of a Skew-Gaussian function with amplitude A, location λ, scale ω, and shape ξ at x.

skewgaussian(x, p)

Calculate the value of a Skew-Gaussian function with parameter vector p at x.

ExtraStats.skewnormal — Function

skewnormal(x, λ, ω, ξ)

Calculate the value of a Skew-Normal function with location λ, scale ω, and shape ξ at x.

skewnormal(x, p)

Calculate the value of a Skew-Normal function with parameter vector p at x.

ExtraStats.SkewNormalCDF — Function

SkewNormalCDF(x, λ, ω, ξ)

Calculate the CDF of a Skew-Normal pdf with location λ, scale ω, and shape ξ at x.

ExtraStats.OwensTFunction — Function

OwensTFunction(h, a)

Calculate the value of Owen's T function with parameters h and a.

ExtraStats.skew_δ — Function

skew_δ(ξ)

Calculate the convienence function δ from shape ξ.

ExtraStats.skew_mean — Function

skew_mean(λ, ω, ξ)

Calculate the mean of a Skew-Normal function with location λ, scale ω, and shape ξ.

ExtraStats.SkewNormalμ2λ — Function

SkewNormalμ2λ(μ, σ, k)

Calculate the location λ of a Skew-Normal function with mean μ, standard deviation σ, and skewness k.

ExtraStats.skew_σ — Function

skew_σ(ω, ξ)

Calculate the standard deviation of a Skew-Normal function with scale ω and shape ξ.

ExtraStats.SkewNormalσ2ω — Function

SkewNormalσ2ω(μ, σ, k)

Calculate the scale parameter ω of a Skew-Normal function with standard deviation σ and skewness k.

ExtraStats.skew_skew — Function

skew_skew(ξ)

Calculate the skewness of a Skew-Normal function with shape ξ.

ExtraStats.SkewNormalSkewness2Posδ — Function

SkewNormalSkewness2Posδ(k)

Calculate the convienence function δ of a Skew-Normal function with skewness k.

ExtraStats.SkewNormalSkewness2Posξ — Function

SkewNormalSkewness2Posξ(k)

Calculate the shape parameter ξ of a Skew-Normal function with skewness k.

ExtraStats.skew_mode — Function

skew_mode(λ, ω, ξ)

Calculate the mode of a Skew-Normal function with location λ, scale ω, and shape ξ.

Log-Skew-Normal Functions

ExtraStats.logskewnormal — Function

logskewnormal(x, λ, ω, ξ)

Calculate the value of a Log-Skew-Normal function with location λ, scale ω, and shape ξ at x.

logskewnormal(x, p)

Calculate the value of a Log-Skew-Normal function with parameter vector p at x.

ExtraStats.LogSkewNormalCDF — Function

LogSkewNormalCDF(x, λ, ω, ξ)

Calculate the CDF of a Log-Skew-Normal pdf with location λ, scale ω, and shape ξ at x.