Library

All functions exported by this package are listed below.

Library Contents

Library

Library Contents

Index

Functions

Micellaneous

Least Squares Fitting

Statistical Distributions

Skew-Normal Functions

Log-Skew-Normal Functions

Index

ExtraStats.CalculateRMSE

ExtraStats.GLS

ExtraStats.KuiperTest

ExtraStats.LogNormalCDF

ExtraStats.LogSkewNormalCDF

ExtraStats.NormalCDF

ExtraStats.OLS

ExtraStats.OwensTFunction

ExtraStats.SE_alpha

ExtraStats.SE_beta

ExtraStats.SkewNormalCDF

ExtraStats.SkewNormalSkewness2Posδ

ExtraStats.SkewNormalSkewness2Posξ

ExtraStats.SkewNormalμ2λ

ExtraStats.SkewNormalσ2ω

ExtraStats.bincounts

ExtraStats.find_HDI

ExtraStats.find_pdf_mode

ExtraStats.fittest

ExtraStats.gaussian

ExtraStats.linear_uncertainty

ExtraStats.lognormal

ExtraStats.logskewnormal

ExtraStats.moving_average

ExtraStats.nanmean

ExtraStats.nanstd

ExtraStats.nanvar

ExtraStats.normal

ExtraStats.r²

ExtraStats.skew_mean

ExtraStats.skew_mode

ExtraStats.skew_skew

ExtraStats.skew_δ

ExtraStats.skew_σ

ExtraStats.skewgaussian

ExtraStats.skewnormal

ExtraStats.sqcovar

ExtraStats.transform_pdf

ExtraStats.uncertainty_lines

ExtraStats.χ²Test

ExtraStats.∫Normaldx

Functions

Micellaneous

ExtraStats.nanmean — Function

nanmean(x; dims = nothing)

Calculate the mean of array x while ignoring NaNs.

ExtraStats.nanvar — Function

nanvar(x; dims = nothing)

Calculate the variance of array x while ignoring NaNs.

ExtraStats.nanstd — Function

nanstd(x; dims = nothing)

Calculate the standard deviation of array x while ignoring NaNs.

ExtraStats.moving_average — Function

moving_average(data, n; replacenans=false)

Calculate the moving average of data with window n.

Optionally replace the NaNs at the start and end with the nearest calculated value if replacenans is true.

ExtraStats.bincounts — Function

bincounts(datax, binedgesx; makepdf=true)

Bin datax in 1 dimension between binedgesx.

Optionally normalize by bin area and number of data points to return a probability density function using makepdf=true.

bincounts(datax, datay, binedgesx, binedgesy; makepdf=true)

Impose a 2d grid on the x-y pairs of data in datax and datay and count the number of points contained in each cell.

Optionally normalize by bin area and number of data points to return a probability density function using makepdf=true.

Least Squares Fitting

ExtraStats.fittest — Function

fittest(x, y; fittype = :OLS, fitmodel = nothing, p0 = [0.1, 0.1], figshow = true)

Perform a least squares fit with dependant variable y and independant x, test the significance of the trend, and create a plot to evaluate the residuals.

Default fitmodel is linear, and fittype ordinary least squares fit, :OLS, but optionally perform generalized least squares with fittype = :GLS.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> fittest(x, y; fittype = :OLS, fitmodel = fmodel, p0 = [0.1, 0.1], figshow = true)
(LsqFit.LsqFitResult{Vector{Float64}, [...]}
------   --------------   --------------
, Float64[]), Test for nonzero correlation
----------------------------
Population details:
    parameter of interest:   Correlation
    value under h_0:         0.0
    point estimate:          0.0611079
    95% confidence interval: (-0.0008886, 0.1226)

Test summary:
    outcome with 95% confidence: fail to reject h_0
    two-sided p-value:           0.0534

Details:
    number of observations:          1000
    number of conditional variables: 0
    t-statistic:                     1.93408
    degrees of freedom:              998
, Scene (1200px, 750px):
  0 Plots
  15 Child Scenes:
    ├ Scene (1200px, 750px)
    [...]

ExtraStats.OLS — Function

OLS(x, y; fitmodel = nothing, p0 = [0.1, 0.1])

Perform an ordinary least squares fit with dependant variable y and independant x.

Default fitmodel is linear.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> OLS(x, y; fitmodel = fmodel, p0 = [0.1, 0.1])
LsqFit.LsqFitResult{Vector{Float64}, [...]}

ExtraStats.GLS — Function

GLS(x, y; K = nothing, fitmodel = nothing, p0 = [0.1, 0.1])

Perform a generalized least squares fit with dependant variable y and independant x.

Default fitmodel is linear.

Weight PrecisionMatrix K is estimated from the residuals of OLS by default.

Example

julia> using ExtraStats, Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> x = 0.1:0.1:100
0.1:0.1:100.0

julia> y = randn(1000)
1000-element Vector{Float64}:
  0.9706563288552144
  0.871497852880908
  ⋮
  0.6530055215854716
 -1.2911234902314417

julia> fmodel(t, p) = p[1] * t .+ p[2]
fmodel (generic function with 1 method)

julia> GLS(x, y; fitmodel = fmodel, p0 = [0.1, 0.1])
LsqFit.LsqFitResult{Vector{Float64}, [...]}

ExtraStats.sqcovar — Function

sqcovar(X)

Calculate the Hermitian square root auto covariance matrix of input vector X.

Fit Evaluation

ExtraStats.CalculateRMSE — Function

CalculateRMSE(e)

Calculate the Root-Mean-Square-Error of the vector of residuals e.

ExtraStats.r² — Function

r²(data, e)

Estimate the coefficient of determination of a fit to data with residuals e.

ExtraStats.χ²Test — Function

χ²Test(x, CDF, p; nbins=500, siglevel = 0.99)

Perform a χ² goodness-of-fit test of the null hypothesis that the data x come from a distribution with cumulative distribution function cdf with parameters p.

Optionally specify the number of bins nbins to separate x into, and the significance level siglevel to test against.

ExtraStats.KuiperTest — Function

KuiperTest(x, Fₕ)

Perform Kuiper's goodness-of-fit test of the null hypothesis that the data x come from CDF Fₕ.

Linear Fit Uncertainty

ExtraStats.linear_uncertainty — Function

linear_uncertainty(xi, yi, yhat, γ)

Calculate the confidence intervals at the γ level on the fit parameters of a linear regression on indpendent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.SE_beta — Function

SE_beta(xi, yi, yhat)

Calculate the standard error of the slope of a linear regression with independent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.SE_alpha — Function

SE_alpha(xi, yi, yhat)

Calculate the standard error of the intercept of a linear regression with independent observations xi, dependent observations yi, and predicted dependent values yhat.

ExtraStats.uncertainty_lines — Function

uncertainty_lines(β_est, α_est, x, t_val, xi, yi, yhat)

Create the upper and lower confidence bound lines of a linear regression.

...

Arguments

β_est: best fit estimate of the slope.

α_est: best fit estimate of the intercept.

x: independent variable values to calculate bounds at.

xi: independent variable observations.

yi: dependent variable observations.

yhat: predicted values of the dependent variable.

γ: confidence level (i.e. γ-th percentile).

...

Statistical Distributions

ExtraStats.find_pdf_mode — Function

find_pdf_mode(pdf, x0; lr=nothing, up=nothing)

Estimate the mode of a probability density function pdf with initial guess x0.

ExtraStats.find_HDI — Function

find_HDI(pdf, samples; target_mass = 0.68)

Find the Highest Density Interval (Hyndman 1996) for the analytic probability density function pdf.

ExtraStats.transform_pdf — Function

transform_pdf(pdf, g, g⁻¹)

Complete a change of variables on an analytic probability density function pdf using the inverse g⁻¹ of the transformation function.

Normal Functions

ExtraStats.gaussian — Function

gaussian(x, A, μ, σ)

Calculate the value of a Gaussian function with amplitude A, centre μ, and standard deviation σ at x.

gaussian(x, p)

Calculate the value of a Gaussian function with parameter vector p at x.

ExtraStats.normal — Function

normal(x, μ, σ)

Calculate the value of a Normal function with centre μ and standard deviation σ at x.

normal(x, p)

Calculate the value of a Normal function with parameter vector p at x.

ExtraStats.∫Normaldx — Function

∫Normaldx(x)

Calculate the value of the integral of a Normal function at x.

ExtraStats.NormalCDF — Function

NormalCDF(x, μ, σ)

Calculate the CDF of a Normal pdf with centre μ and standard deviation σ at x.

Log-Normal Functions

ExtraStats.lognormal — Function

lognormal(x, μ, σ)

Calculate the value of a Log-Normal function with log-centre μ and log-standard deviation σ at x.

lognormal(x, p)

Calculate the value of a Log-Normal function with parameter vector p at x.

ExtraStats.LogNormalCDF — Function

LogNormalCDF(x, μ, σ)

Calculate the CDF of a Log-Normal pdf with centre μ and standard deviation σ at x.

Skew-Normal Functions

ExtraStats.skewgaussian — Function

skewgaussian(x, A, λ, ω, ξ)

Calculate the value of a Skew-Gaussian function with amplitude A, location λ, scale ω, and shape ξ at x.

skewgaussian(x, p)

Calculate the value of a Skew-Gaussian function with parameter vector p at x.

ExtraStats.skewnormal — Function

skewnormal(x, λ, ω, ξ)

Calculate the value of a Skew-Normal function with location λ, scale ω, and shape ξ at x.

skewnormal(x, p)

Calculate the value of a Skew-Normal function with parameter vector p at x.

ExtraStats.SkewNormalCDF — Function

SkewNormalCDF(x, λ, ω, ξ)

Calculate the CDF of a Skew-Normal pdf with location λ, scale ω, and shape ξ at x.

ExtraStats.OwensTFunction — Function

OwensTFunction(h, a)

Calculate the value of Owen's T function with parameters h and a.

ExtraStats.skew_δ — Function

skew_δ(ξ)

Calculate the convienence function δ from shape ξ.

ExtraStats.skew_mean — Function

skew_mean(λ, ω, ξ)

Calculate the mean of a Skew-Normal function with location λ, scale ω, and shape ξ.

ExtraStats.SkewNormalμ2λ — Function

SkewNormalμ2λ(μ, σ, k)

Calculate the location λ of a Skew-Normal function with mean μ, standard deviation σ, and skewness k.

ExtraStats.skew_σ — Function

skew_σ(ω, ξ)

Calculate the standard deviation of a Skew-Normal function with scale ω and shape ξ.

ExtraStats.SkewNormalσ2ω — Function

SkewNormalσ2ω(μ, σ, k)

Calculate the scale parameter ω of a Skew-Normal function with standard deviation σ and skewness k.

ExtraStats.skew_skew — Function

skew_skew(ξ)

Calculate the skewness of a Skew-Normal function with shape ξ.

ExtraStats.SkewNormalSkewness2Posδ — Function

SkewNormalSkewness2Posδ(k)

Calculate the convienence function δ of a Skew-Normal function with skewness k.

ExtraStats.SkewNormalSkewness2Posξ — Function

SkewNormalSkewness2Posξ(k)

Calculate the shape parameter ξ of a Skew-Normal function with skewness k.

ExtraStats.skew_mode — Function

skew_mode(λ, ω, ξ)

Calculate the mode of a Skew-Normal function with location λ, scale ω, and shape ξ.

Log-Skew-Normal Functions

ExtraStats.logskewnormal — Function

logskewnormal(x, λ, ω, ξ)

Calculate the value of a Log-Skew-Normal function with location λ, scale ω, and shape ξ at x.

logskewnormal(x, p)

Calculate the value of a Log-Skew-Normal function with parameter vector p at x.

ExtraStats.LogSkewNormalCDF — Function

LogSkewNormalCDF(x, λ, ω, ξ)

Calculate the CDF of a Log-Skew-Normal pdf with location λ, scale ω, and shape ξ at x.