Miscellaneous

These functions are for general use, and can come in handy in a variety of situations.

ExtraStats.nanmean — Function

nanmean(x; dims = nothing)

Calculate the mean of array x while ignoring NaNs.

ExtraStats.nanvar — Function

nanvar(x; dims = nothing)

Calculate the variance of array x while ignoring NaNs.

ExtraStats.nanstd — Function

nanstd(x; dims = nothing)

Calculate the standard deviation of array x while ignoring NaNs.

ExtraStats.moving_average — Function

moving_average(data, n; replacenans=false)

Calculate the moving average of data with window n.

Optionally replace the NaNs at the start and end with the nearest calculated value if replacenans is true.

ExtraStats.bincounts — Function

bincounts(datax, binedgesx; makepdf=true)

Bin datax in 1 dimension between binedgesx.

Optionally normalize by bin area and number of data points to return a probability density function using makepdf=true.

bincounts(datax, datay, binedgesx, binedgesy; makepdf=true)

Impose a 2d grid on the x-y pairs of data in datax and datay and count the number of points contained in each cell.

Optionally normalize by bin area and number of data points to return a probability density function using makepdf=true.